[#95] Fix context event dispatch
